Component storage loader directive modal update
Close event action for "x" in modal

this is a small change to catch the close event and close the modal when someone clicks on the x that was implemented in the style-guide corrections for the storage-client

added Attributes to restrict property of directive

This makes it more extensible by using attributes

added useCtrl to scope to bind controller

Binds controller to whatever is passed in. read only. for greater extensibility

added attrs to link and changed storageUrl

This takes companyid from attr that is passed in.

clean up removal for code that is not needed anymore

added a controller and instance modal in directive

moved the $window addEventListener inside the scope.open function to enable 'x' button to close modal.  updated module dependency to have ui.bootstrap and the $modal svc. cleaned up spacing of code

got rid of unneeded code to click backdrop

Put in new code to dismiss the modalInstance when the close event is fired.

dismiss doesn't need a value

corrected this mistake.

made changes to the loader.html

to make use of the ui-bootstrap modal and give correct css for the modal.

add the ui-bootstrap dependency

added new radio button for new directive usage

created a new radio button to explain the custom and extensibility of the new directive.

ran gulp build to update dist folder