attribute local added to selector
chgs to have style-guide use local files

changed the demo file to use custom script that has new local attribute that tells the directive to load the local json for files (for example) or to actually view the contents of a company id.  The demo should use the example. Also added verbiage to let user know this.

ran gulp build to update dist